refactor: extract git remote parsing helpers
